Extract AbstractAsyncCmmnHistoryManager that contains common methods for populating history data
